@@ -5,43 +5,45 @@
 	"version": "1.0.0",
 	"author": "Flexera, Dramba Victor",
 	"dependencies": {
-		"@types/styled-components": "^5.1.2",
-		"copy-webpack-plugin": "^9.0.1",
+		"@types/styled-components": "^5.1.15",
+		"clean-webpack-plugin": "^4.0.0",
+		"copy-webpack-plugin": "^9.1.0",
 		"dagre": "^0.8.5",
-		"file-loader": "6.0.0",
-		"react": "16.12.0",
-		"react-dom": "16.12.0",
-		"react-is": "16.12.0",
-		"react-router": "5.1.2",
-		"react-router-dom": "5.1.2",
-		"styled-components": "^5.1.1",
-		"svg-editor-tools": "github:dvictor/svg-editor-tools#1.0.1"
+		"file-loader": "6.2.0",
+		"react": "17.0.2",
+		"react-dom": "17.0.2",
+		"react-is": "17.0.2",
+		"react-router": "6.0.2",
+		"react-router-dom": "6.0.2",
+		"styled-components": "^5.3.3",
+		"svg-editor-tools": "github:dvictor/svg-editor-tools#1.0.1",
+		"yarn": "^1.22.17"
 	},
 	"devDependencies": {
-		"@babel/core": "7.9.6",
-		"@babel/preset-env": "7.9.6",
-		"@babel/preset-react": "7.9.4",
-		"@babel/preset-typescript": "7.9.0",
-		"@pmmmwh/react-refresh-webpack-plugin": "0.3.1",
-		"@types/dagre": "^0.7.44",
-		"@types/react": "16.9.23",
-		"@types/react-dom": "16.9.5",
-		"@types/react-router": "5.1.4",
-		"@types/react-router-dom": "5.1.3",
-		"@types/react-syntax-highlighter": "11.0.4",
-		"babel-loader": "8.0.6",
-		"css-loader": "3.4.2",
-		"eslint": "6.8.0",
-		"eslint-plugin-react": "7.18.3",
-		"eslint-plugin-react-hooks": "2.4.0",
-		"html-loader": "0.5.5",
-		"html-webpack-plugin": "3.2.0",
-		"react-refresh": "0.8.2",
-		"style-loader": "1.1.3",
-		"typescript": "3.8.2",
-		"webpack": "4.41.6",
-		"webpack-cli": "3.3.11",
-		"webpack-dev-server": "^3.11.2"
+		"@babel/core": "7.16.0",
+		"@babel/preset-env": "7.16.0",
+		"@babel/preset-react": "7.16.0",
+		"@babel/preset-typescript": "7.16.0",
+		"@pmmmwh/react-refresh-webpack-plugin": "^0.5.1",
+		"@types/dagre": "^0.7.46",
+		"@types/react": "17.0.34",
+		"@types/react-dom": "17.0.11",
+		"@types/react-router": "5.1.17",
+		"@types/react-router-dom": "5.3.2",
+		"@types/react-syntax-highlighter": "13.5.2",
+		"babel-loader": "8.2.3",
+		"css-loader": "6.5.1",
+		"eslint": "^8.2.0",
+		"eslint-plugin-react": "7.27.0",
+		"eslint-plugin-react-hooks": "4.3.0",
+		"html-loader": "3.0.1",
+		"html-webpack-plugin": "5.5.0",
+		"react-refresh": "0.10.0",
+		"style-loader": "3.3.1",
+		"typescript": "4.4.4",
+		"webpack": "5.64.0",
+		"webpack-cli": "^4.9.1",
+		"webpack-dev-server": "^4.5.0"
 	},
 	"scripts": {
 		"build": "NODE_ENV=production webpack",
