@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 /* eslint-env node */
 const HtmlWebPackPlugin = require('html-webpack-plugin');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const path = require('path');
 
 const base = require('./webpack.config.base')
@@ -11,11 +12,14 @@ base.plugins.push(
 		template: './src/static/index.html',
 		filename: './index.html',
 	}),
+	new CleanWebpackPlugin({
+		protectWebpackAssets: false,
+		cleanAfterEveryBuildPatterns: ['*.LICENSE.txt'],
+	})
 )
 base.output = {
 	path: path.resolve(__dirname, 'dist-static/'),
 	publicPath: '.',
 }
-base.devtool = '' // no sourcemap
 
 module.exports = base
